fix: match cdf_GDR_t's integer fields to the CDF spec's signed int4 type
Apple Clang rejected a macOS CI wheel build with a hard narrowing-conversion
error at records-loading.hpp:98 (vdr_context_t<version_t>{ gdr.rNumDims }):
GCC only warns on narrowing uint32_t -> int32_t in a brace-init list, Clang
enforces the standard's ill-formed rule.

Checked against the actual spec (NASA CDF Internal Format Description v3.9)
rather than just cross-matching CDFpp's own, inconsistently-typed structs:
every GDR field is documented as a "Signed 4-byte integer" -- there is no
unsigned integer type anywhere in CDF's internal record format (UInt4/2/1
data types exist only for variable *values*, never record-header fields).
cdf_GDR_t::rNumDims (and NrVars, NumAttr, rMaxRec, NzVars,
LeapSecondLastUpdated, rDimSizes) were declared uint32_t, a pre-existing
mismatch predating this migration -- fixed all seven to int32_t, matching
the spec and vdr_context_t's already-correct type (left unchanged).

Zero on-disk difference for real files (values are always small and
non-negative, same bytes either way). Fixed two knock-on narrowing sites
this surfaced: loading.hpp's `NzVars + NrVars` sum into a size_t brace-init
(now int->size_t, itself narrowing -- explicit static_cast), and
saving.hpp's rMaxRec initializer, which encoded the spec's documented
"-1 means no records" sentinel via the uint32_t bit-pattern 0xFFFFFFFF --
now written directly as -1, matching cdf_GDR_t's declared type instead of
relying on the bit-pattern coincidence.

Verified clean under both GCC and Clang (full rebuild + 27/27 test suite
each), including a knock-on field_size overload in cdf_GDR_t left
declaring the old uint32_t signature.

+++ include/cdfpp/cdf-io/loading/loading.hpp
@@ -83,7 +83,8 @@ namespace
     [[nodiscard]] std::optional<CDF> impl_parse_cdf(
         parsing_context_t& parsing_context, bool lazy_load = false)
     {
-        common::cdf_repr repr { parsing_context.gdr.NzVars + parsing_context.gdr.NrVars };
+        common::cdf_repr repr { static_cast<std::size_t>(
+            parsing_context.gdr.NzVars + parsing_context.gdr.NrVars) };
         repr.majority = parsing_context.majority;
         repr.distribution_version = parsing_context.distribution_version();
         repr.compression_type = parsing_context.compression_type;
